+++ tests/extracted/test_main.py
@@ -9,13 +9,12 @@ from mex.common.models import (
     ExtractedOrganizationalUnit,
 )
 
-if TYPE_CHECKING:  # pragma: no cover
+if TYPE_CHECKING:
     from fastapi.testclient import TestClient
 
     from tests.conftest import DummyData, MockedGraph
 
 
-@pytest.mark.usefixtures("mocked_valkey")
 def test_search_extracted_items_mocked(
     client_with_api_key_read_permission: TestClient,
     mocked_graph: MockedGraph,