+++ tests/merged/test_main.py
@@ -14,13 +14,12 @@ from mex.common.models import (
 )
 from mex.common.types import Validation
 
-if TYPE_CHECKING:  # pragma: no cover
+if TYPE_CHECKING:
     from fastapi.testclient import TestClient
 
     from tests.conftest import DummyData, DummyDataName, MockedGraph
 
 
-@pytest.mark.usefixtures("mocked_valkey")
 def test_search_merged_items_mocked(
     client_with_api_key_read_permission: TestClient,
     mocked_graph: MockedGraph,